+++ src/Bot/Capability/FetchFeed.hs
@@ -1,0 +1,6 @@
+module Bot.Capability.FetchFeed where
+
+import Text.Feed.Types (Feed)
+
+class (Monad m) => FetchFeed m where
+  fetchFeed :: String -> m (Maybe Feed)
